Add memory infrastructure and security fixes to Carta
Shared library additions:
- Add SHArena arena allocator for same-lifetime allocations
- Add SHPool contiguous memory pool with offset-based access
- Add sh_safe_mul_size() overflow-safe multiplication helper
- Add SAFE_FREE macro for post-free hygiene

Carta security fixes:
- Move role hash table to per-context (thread safety)
- Make render cache thread-local (thread safety)
- Add pthread_once for CRC32 table initialization
- Fix division by zero in clipping functions
- Fix uint32_t bounds check wrap-around in role_idx
- Fix undefined behavior in zigzag encoding (left shift)

Carta memory optimizations:
- Add parse arena for PBF temporary arrays
- Add coordinate pool for way geometry
- Add scratch buffers for multipolygon assembly
- Add pre-allocated edge buffers for polygon rendering

Code quality improvements:
- Add named constants for magic numbers
- Add OWNERSHIP documentation to public APIs
- Change ct_cache_get/put return type from int to bool
- Remove unused point_pool from CTTile

Ralph/Fuelwise/Velo:
- Backport shared arena/pool to Ralph (replaces local implementation)
- Update Fuelwise Makefile to link shared library
- Add NOMP=1 support to Velo Makefile for macOS

All 396 tests pass. ASan/UBSan validation clean.

+++ carta/src/ct_mvt.c
@@ -103,8 +103,8 @@ static void enc_write_varint(CTMVTEncoder *enc, uint64_t value)
 
 static void enc_write_svarint(CTMVTEncoder *enc, int64_t value)
 {
-    /* Zigzag encode */
-    uint64_t uval = (uint64_t)((value << 1) ^ (value >> 63));
+    /* Zigzag encode: cast to unsigned before left shift to avoid UB */
+    uint64_t uval = ((uint64_t)value << 1) ^ (uint64_t)(value >> 63);
     enc_write_varint(enc, uval);
 }
